removed external references to WeightedPairDistribution
